+++ R/RcppExports.R
@@ -17,3 +17,7 @@ best_merge_cpp <- function(pm, nvec, lambda, k) {
     .Call(`_stagedtrees_best_merge_cpp`, pm, nvec, lambda, k)
 }
 
+best_move_cpp <- function(ct, asg, nstage, lambda) {
+    .Call(`_stagedtrees_best_move_cpp`, ct, asg, nstage, lambda)
+}
+